Rearranged the download urls/views/tests
